ajout de la reconnaissance du rythme de la musique pour la map 2

+++ OedivXuej/Source/OedivXuej/MusicAnalysis.h
@@ -9,14 +9,14 @@
 #include "MusicAnalysis.generated.h"
 
 /**
- * 
+ *
  */
 UCLASS()
 class OEDIVXUEJ_API UMusicAnalysis : public UBlueprintFunctionLibrary
 {
 	GENERATED_BODY()
-	
+
 public:
 	UFUNCTION(BlueprintCallable, Category = "MusicAnalysis")
-		static int GetBpm(FString SourcePath);
+		static float GetBpm(FString SourcePath);
 };
